pyproject.toml
[project]
name = "notiq"
version = "0.0.1"
description = "Notification, analytics and job queue system. currently in planning phase"
readme = "README.md"
authors = [
    { name = "HERITAGE-XION", email = "iheritage934@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "celery-redbeat>=2.3.3",
    "celery[redis]>=5.6.2",
    "prometheus-client>=0.24.1",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.31",
    "tzlocal>=5.3.1",
]

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "ruff>=0.14.13",
]


# --- PYRIGHT CONFIGURATION ---
[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "strict"
reportMissingImports = true
reportUnknownMemberType = false
reportMissingTypeStubs = false
reportUnnecessaryCast = false
reportUnknownVariableType = false
exclude = ["**/tests/**", "**/migrations/**"]


# --- RUFF CONFIGURATION (Linter & Formatter) ---
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
# E/W = pycodestyle (standard errors)
# F = pyflakes (unused imports, variables)
# I = isort (import sorting)
# B = flake8-bugbear (common bugs)
# UP = pyupgrade (modernize syntax)
select = ["E", "W", "F", "I", "B", "UP", "C4"]
ignore = [
    "F401",  # unused import (Pyright covers this)
    "F841",  # unused variable (Pyright covers this)
    "B007",  # unused loop control variable
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# --- MYPY CONFIGURATION (Type Checker) ---
[tool.mypy]
strict = true
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
exclude = "^(tests|migrations)/"

# If a library doesn't have type hints, ignore it here:
[[tool.mypy.overrides]]
module = "notiq.utils.dicovery"
disable_error_code = ["attr-defined"]  # Path.walk() is Python 3.12+

[[tool.mypy.overrides]]
module = "notiq.config"
disable_error_code = ["misc"]  # pydantic-settings BaseSettings typing issue
